# packages/nextjs/public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lunargistics Mission Planning Suite API
  description: |
    REST API for the Lunargistics Mission Planning Suite, providing endpoints for mission management,
    orbital calculations, document generation, and blockchain integration.
    
    ## Authentication
    All endpoints require authentication via Bearer token in the Authorization header.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limit headers are included in all responses.
  version: 1.0.0
  contact:
    name: Lunargistics Development Team
    email: api@lunargistics.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.lunargistics.com/v1
    description: Production server
  - url: https://staging-api.lunargistics.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Mission:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: string
          description: Unique mission identifier
          example: "mission-123"
        name:
          type: string
          description: Mission name
          example: "Lunar Gateway Resupply"
        type:
          type: string
          enum: [Earth Observation, Communication, Navigation, Scientific, Exploration]
          description: Mission type
        description:
          type: string
          description: Detailed mission description
        status:
          type: string
          enum: [Planning, Active, Complete, Cancelled]
          default: Planning
        budget:
          type: object
          properties:
            total:
              type: number
              description: Total budget in USD
            allocated:
              type: number
              description: Allocated budget
            spent:
              type: number
              description: Spent budget
        phase:
          type: string
          enum: [Pre-Phase A, Phase A, Phase B, Phase C, Phase D, Phase E, Phase F]
          description: Mission lifecycle phase
        ipfsHash:
          type: string
          description: IPFS hash for mission data
          example: "QmXoypizjW3WknFiJnKLwHCnL72vedxjQkDDP1mXWo6uco"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          description: Rate limit error message
        limit:
          type: integer
          description: Request limit
        remaining:
          type: integer
          description: Remaining requests
        reset:
          type: string
          format: date-time
          description: Rate limit reset time
  responses:
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
          description: Rate limit reset time
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit reset
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
security:
  - bearerAuth: []
paths:
  /missions:
    get:
      tags:
        - Missions
      summary: List all missions
      description: Retrieve a list of all missions with optional filtering
      parameters:
        - in: query
          name: id
          schema:
            type: string
          description: Filter by mission ID
        - in: query
          name: name
          schema:
            type: string
          description: Filter by mission name
        - in: query
          name: type
          schema:
            type: string
            enum: [Earth Observation, Communication, Navigation, Scientific, Exploration]
          description: Filter by mission type
        - in: query
          name: status
          schema:
            type: string
            enum: [Planning, Active, Complete, Cancelled]
          description: Filter by mission status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  missions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mission'
                  total:
                    type: integer
                    description: Total number of missions
                  page:
                    type: integer
                    description: Current page
                  pages:
                    type: integer
                    description: Total pages
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    post:
      tags:
        - Missions
      summary: Create a new mission
      description: Create a new mission and store it on IPFS and blockchain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mission'
      responses:
        '201':
          description: Mission created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  ipfsHash:
                    type: string
                  message:
                    type: string
                  mission:
                    $ref: '#/components/schemas/Mission'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...